Moving CLI helpers to their own section not imported by default.
